+++ capstone-sys/build.rs
@@ -40,13 +40,22 @@ extern crate bindgen;
 
 extern crate cc;
 
+#[cfg(feature = "use_bindgen")]
+extern crate regex;
+
+#[cfg(feature = "use_bindgen")]
+use {
+    regex::Regex,
+    std::{fs::File, io::Write},
+};
+
 use std::env;
 use std::fs::copy;
 use std::path::PathBuf;
 
 include!("common.rs");
 
-const CAPSTONE_DIR: &'static str = "capstone";
+const CAPSTONE_DIR: &str = "capstone";
 
 /// Indicates how capstone library should be linked
 #[allow(dead_code)]
@@ -79,7 +88,8 @@ fn build_capstone_cc() {
 
     fn find_c_source_files(dir: &str) -> Vec<String> {
         read_dir_and_filter(dir, |e| {
-            let file_type = e.file_type()
+            let file_type = e
+                .file_type()
                 .expect("Failed to read capstone source directory");
             let file_name = e.file_name().into_string().expect("Invalid filename");
             file_type.is_file() && (file_name.ends_with(".c") || file_name.ends_with(".C"))
@@ -88,7 +98,8 @@ fn build_capstone_cc() {
 
     fn find_arch_dirs() -> Vec<String> {
         read_dir_and_filter(&format!("{}/{}", CAPSTONE_DIR, "arch"), |e| {
-            let file_type = e.file_type()
+            let file_type = e
+                .file_type()
                 .expect("Failed to read capstone source directory");
             file_type.is_dir()
         })
@@ -100,8 +111,8 @@ fn build_capstone_cc() {
     }
 
     let use_static_crt = {
-        let target_features = env::var("CARGO_CFG_TARGET_FEATURE").unwrap_or(String::new());
-        target_features.split(",").any(|f| f == "crt-static")
+        let target_features = env::var("CARGO_CFG_TARGET_FEATURE").unwrap_or_default();
+        target_features.split(',').any(|f| f == "crt-static")
     };
 
     cc::Build::new()
@@ -115,9 +126,11 @@ fn build_capstone_cc() {
         .define("CAPSTONE_HAS_M68K", None)
         .define("CAPSTONE_HAS_MIPS", None)
         .define("CAPSTONE_HAS_POWERPC", None)
+        .define("CAPSTONE_HAS_RISCV", None)
         .define("CAPSTONE_HAS_SPARC", None)
         .define("CAPSTONE_HAS_SYSZ", None)
         .define("CAPSTONE_HAS_TMS320C64X", None)
+        .define("CAPSTONE_HAS_WASM", None)
         .define("CAPSTONE_HAS_X86", None)
         .define("CAPSTONE_HAS_XCORE", None)
         .flag_if_supported("-Wno-unused-function")
@@ -133,7 +146,7 @@ fn build_capstone_cc() {
 
 /// Search for header in search paths
 #[cfg(feature = "use_bindgen")]
-fn find_capstone_header(header_search_paths: &Vec<PathBuf>, name: &str) -> Option<PathBuf> {
+fn find_capstone_header(header_search_paths: &[PathBuf], name: &str) -> Option<PathBuf> {
     for search_path in header_search_paths.iter() {
         let potential_file = search_path.join(name);
         if potential_file.is_file() {
@@ -145,19 +158,89 @@ fn find_capstone_header(header_search_paths: &Vec<PathBuf>, name: &str) -> Optio
 
 /// Gets environment variable value. Panics if variable is not set.
 fn env_var(var: &str) -> String {
-    env::var(var).expect(&format!("Environment variable {} is not set", var))
+    env::var(var).unwrap_or_else(|_| panic!("Environment variable {} is not set", var))
+}
+
+/// Parse generated bindings and impl from_insn_id() for all architectures
+/// instructions enum declaration.
+#[cfg(feature = "use_bindgen")]
+fn impl_insid_to_insenum(bindings: &str) -> String {
+    let mut impl_arch_enum = String::new();
+    impl_arch_enum.push_str("use core::convert::From;\n");
+
+    for cs_arch in ARCH_INCLUDES {
+        let arch = cs_arch.cs_name();
+
+        // find architecture instructions enum declaration
+        let re_enum_def = Regex::new(&format!("pub enum {}_insn (?s)\\{{.*?\\}}", arch))
+            .expect("Unable to compile regex");
+        let cap_enum_def = &re_enum_def
+            .captures(bindings)
+            .expect("Unable to capture group")[0];
+
+        // find instructions and their id
+        let re_ins_ids = Regex::new(&format!(
+            "{}_INS_(?P<ins>[A-Z0-9_]+) = (?P<id>\\d+)",
+            &arch.to_uppercase()
+        ))
+        .expect("Unable to compile regex");
+
+        impl_arch_enum.push_str(&format!(
+            "impl From<u32> for {}_insn {{\n
+            fn from(id: u32) -> Self {{\n
+            match id {{\n",
+            &arch
+        ));
+
+        // fill match expression
+        for cap_ins_id in re_ins_ids.captures_iter(cap_enum_def) {
+            impl_arch_enum.push_str(&format!(
+                "{} => {}_insn::{}_INS_{},\n",
+                &cap_ins_id["id"],
+                &arch,
+                &arch.to_uppercase(),
+                &cap_ins_id["ins"]
+            ));
+        }
+
+        // if id didn't match, return [arch]_INS_INVALID.
+        // special case for m680x which has 'INVLD' variant instead of 'INVALID'
+        match arch {
+            "m680x" => {
+                impl_arch_enum.push_str(&format!(
+                    "_ => {}_insn::{}_INS_INVLD,",
+                    &arch,
+                    &arch.to_uppercase()
+                ));
+            }
+            _ => {
+                impl_arch_enum.push_str(&format!(
+                    "_ => {}_insn::{}_INS_INVALID,",
+                    &arch,
+                    &arch.to_uppercase()
+                ));
+            }
+        }
+
+        impl_arch_enum.push_str("}\n}\n}\n");
+    }
+
+    impl_arch_enum
 }
 
 /// Create bindings using bindgen
 #[cfg(feature = "use_bindgen")]
 fn write_bindgen_bindings(
-    header_search_paths: &Vec<PathBuf>,
+    header_search_paths: &[PathBuf],
     update_pregenerated_bindings: bool,
     pregenerated_bindgen_header: PathBuf,
-    out_path: PathBuf,
+    pregenerated_bindgen_impl: PathBuf,
+    out_bindings_path: PathBuf,
+    out_impl_path: PathBuf,
 ) {
     let mut builder = bindgen::Builder::default()
         .rust_target(bindgen::RustTarget::Stable_1_19)
+        .size_t_is_usize(true)
         .use_core()
         .ctypes_prefix("libc")
         .header(
@@ -172,14 +255,14 @@ fn write_bindgen_bindings(
         .layout_tests(false) // eliminate test failures on platforms with different pointer sizes
         .impl_debug(true)
         .constified_enum_module("cs_err|cs_group_type|cs_opt_value")
-        .bitfield_enum("cs_mode")
+        .bitfield_enum("cs_mode|cs_ac_type")
         .rustified_enum(".*");
 
     // Whitelist cs_.* functions and types
     let pattern = String::from("cs_.*");
     builder = builder
-        .whitelist_function(&pattern)
-        .whitelist_type(&pattern);
+        .allowlist_function(&pattern)
+        .allowlist_type(&pattern);
 
     // Whitelist types with architectures
     for arch in ARCH_INCLUDES {
@@ -187,7 +270,7 @@ fn write_bindgen_bindings(
         let arch_type_pattern = format!(".*(^|_){}(_|$).*", arch.cs_name);
         let const_mod_pattern = format!("^{}_(reg|insn_group)$", arch.cs_name);
         builder = builder
-            .whitelist_type(&arch_type_pattern)
+            .allowlist_type(&arch_type_pattern)
             .constified_enum_module(&const_mod_pattern);
     }
 
@@ -195,11 +278,20 @@ fn write_bindgen_bindings(
 
     // Write bindings to $OUT_DIR/bindings.rs
     bindings
-        .write_to_file(out_path.clone())
+        .write_to_file(&out_bindings_path)
         .expect("Unable to write bindings");
 
+    // Parse bindings and impl fn to cast u32 to <arch>_insn, write output to file
+    let bindings_impl_str = impl_insid_to_insenum(&bindings.to_string());
+    let mut bindings_impl = File::create(&out_impl_path).expect("Unable to open file");
+    bindings_impl
+        .write_all(bindings_impl_str.as_bytes())
+        .expect("Unable to write file");
+
     if update_pregenerated_bindings {
-        copy(out_path, pregenerated_bindgen_header).expect("Unable to update capstone bindings");
+        copy(out_bindings_path, pregenerated_bindgen_header)
+            .expect("Unable to update capstone bindings");
+        copy(out_impl_path, pregenerated_bindgen_impl).expect("Unable to update capstone bindings");
     }
 }
 
@@ -226,23 +318,27 @@ fn main() {
 
     // If UPDATE_CAPSTONE_BINDINGS is set, then updated the pre-generated capstone bindings
     let update_pregenerated_bindings = env::var("UPDATE_CAPSTONE_BINDINGS").is_ok();
-    if update_pregenerated_bindings {
-        assert!(
-            cfg!(feature = "use_bindgen"),
-            concat!(
-                "Setting UPDATE_CAPSTONE_BINDINGS only makes ",
-                "sense when enabling feature \"use_bindgen\""
-            )
-        );
+    if update_pregenerated_bindings && !cfg!(feature = "use_bindgen") {
+        panic!( "Setting UPDATE_CAPSTONE_BINDINGS only makes sense when enabling feature \"use_bindgen\"");
     }
 
     let pregenerated_bindgen_header: PathBuf = [
         env_var("CARGO_MANIFEST_DIR"),
         "pre_generated".into(),
         BINDINGS_FILE.into(),
-    ].iter()
-        .collect();
-    let out_path = PathBuf::from(env_var("OUT_DIR")).join(BINDINGS_FILE);
+    ]
+    .iter()
+    .collect();
+    let pregenerated_bindgen_impl: PathBuf = [
+        env_var("CARGO_MANIFEST_DIR"),
+        "pre_generated".into(),
+        BINDINGS_IMPL_FILE.into(),
+    ]
+    .iter()
+    .collect();
+
+    let out_bindings_path = PathBuf::from(env_var("OUT_DIR")).join(BINDINGS_FILE);
+    let out_impl_path = PathBuf::from(env_var("OUT_DIR")).join(BINDINGS_IMPL_FILE);
 
     // Only run bindgen if we are *not* using the bundled capstone bindings
     #[cfg(feature = "use_bindgen")]
@@ -250,10 +346,17 @@ fn main() {
         &header_search_paths,
         update_pregenerated_bindings,
         pregenerated_bindgen_header,
-        out_path,
+        pregenerated_bindgen_impl,
+        out_bindings_path,
+        out_impl_path,
     );
 
     // Otherwise, copy the pregenerated bindings
     #[cfg(not(feature = "use_bindgen"))]
-    copy(&pregenerated_bindgen_header, &out_path).expect("Unable to update capstone bindings");
+    {
+        copy(&pregenerated_bindgen_header, &out_bindings_path)
+            .expect("Unable to update capstone bindings");
+        copy(&pregenerated_bindgen_impl, &out_impl_path)
+            .expect("Unable to update capstone bindings");
+    }
 }