+++ CHANGELOG.md
@@ -5,10 +5,48 @@ Notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - YYYY-MM-DD
+## UNRELEASED - YYYY-MM-DD
+### Fixed
+- Document that minimum supported Rust version is 1.50.0
+    - Improperly documented as 1.40.0 in 0.14.0 release
+
+## [0.10.0] - 2021-08-09
+
+### Added
+- Links to related types/functions in API documentation
+
+### Changed
+- Access `Insn` elements of `Instructions` via `Deref`/`AsRef` traits instead of an iterator
+- Access `Insn` groups and read/write registers via a slice
+- Update README.md code example
+
+## [0.9.0] - 2021-07-13
+
+### Added
+- Support for RISC-V architecture
+
+### Changed
+- Updated minimum supported Rust version to 1.40.0
+
+## [0.8.0] - 2021-04-09
+
+### Added
+- `Insn::from_raw()` to convert from a `cs_insn` pointer
+- `Deref` impl for `Instructions`
+
+### Changed
+- X86: return `X86OpMem::segment()` as `RegId` instead of `u32`
+
+### Fixed
+- Fixed data race in multi-threaded environment ([#87](https://github.com/capstone-rust/capstone-rs/issues/87))
+
+## [0.7.0] - 2020-03-16
 
 ### Added
 - `no_std` compatibility
+- Parallel disassemble example
+- X86: add `X86Operand.access` field
+- Implement `From<u32>` for *_ins enums
 
 ### Changed
 - Bump minimum Rust version to 1.36.0
@@ -109,7 +147,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Dependency
 
-[Unreleased]: https://github.com/capstone-rust/capstone-rs/compare/capstone-v0.6.0...master
+[0.10.0]: https://github.com/capstone-rust/capstone-rs/compare/capstone-v0.9.0...capstone-v0.10.0
+[0.9.0]: https://github.com/capstone-rust/capstone-rs/compare/capstone-v0.8.0...capstone-v0.9.0
+[0.8.0]: https://github.com/capstone-rust/capstone-rs/compare/capstone-v0.7.0...capstone-v0.8.0
+[0.7.0]: https://github.com/capstone-rust/capstone-rs/compare/capstone-v0.6.0...capstone-v0.7.0
 [0.6.0]: https://github.com/capstone-rust/capstone-rs/compare/v0.5.0...capstone-v0.6.0
 [0.5.0]: https://github.com/capstone-rust/capstone-rs/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/capstone-rust/capstone-rs/compare/v0.3.1...v0.4.0